4.开发/front/demo/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
cryptography==42.0.5
curl_cffi==0.6.1
dataclasses-json==0.6.4
distro==1.9.0
duckduckgo_search==4.5.0
etelemetry==0.3.1
exceptiongroup==1.2.0
extra-streamlit-components==0.1.70
filelock==3.13.1
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langsmith==0.1.7
loguru==0.7.2
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
playwright==1.41.2
prettytable==3.10.0
protobuf==4.25.3
prov==2.0.0
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
pyecharts==2.0.4
pyee==11.0.1
PyExecJS==1.5.1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pyparsing==3.1.1
PyPDF2==3.0.1
pyproject==1.3.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
scipy==1.12.0
selenium==4.18.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
starlette==0.37.1
streamlit==1.31.1
streamlit-authenticator==0.3.1
streamlit-cookies-manager==0.2.0
streamlit-echarts==0.4.0
streamlit-feedback==0.1.3
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traits==6.3.2
trio==0.24.0
trio-websocket==0.11.1
trubrics==1.6.2
trycourier==5.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
